requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.2
Flask-Bcrypt==1.0.1
yfinance==0.2.20
numpy==1.24.3
pandas==1.5.3
scikit-learn==1.2.2
PyJWT==2.6.0